test: fail fast on the integration test

+++ redis_test.go
@@ -44,13 +44,13 @@ func testFlush(t *testing.T, client Client) {
 
 	kvs := make(map[string]string, keys)
 	for i := 0; i < keys; i++ {
-		kvs[strconv.Itoa(i)] = strconv.FormatInt(rand.Int63(), 10)
+		kvs["f"+strconv.Itoa(i)] = strconv.FormatInt(rand.Int63(), 10)
 	}
 
 	t.Logf("prepare %d keys for FLUSH\n", keys)
 	jobs, wait := parallel(para)
-	for i := 0; i < keys; i++ {
-		key := strconv.Itoa(i)
+	for i := 0; i < keys && !t.Failed(); i++ {
+		key := "f" + strconv.Itoa(i)
 		jobs <- func() {
 			val, err := client.Do(ctx, client.B().Set().Key(key).Value(kvs[key]).Build()).ToString()
 			if err != nil || val != "OK" {
@@ -59,11 +59,14 @@ func testFlush(t *testing.T, client Client) {
 		}
 	}
 	wait()
+	if t.Failed() {
+		return
+	}
 
 	t.Logf("testing client side caching before flush\n")
 	jobs, wait = parallel(para)
-	for i := 0; i < keys; i++ {
-		key := strconv.Itoa(i)
+	for i := 0; i < keys && !t.Failed(); i++ {
+		key := "f" + strconv.Itoa(i)
 		jobs <- func() {
 			resp := client.DoCache(ctx, client.B().Get().Key(key).Cache(), time.Minute)
 			val, err := resp.ToString()
@@ -76,6 +79,9 @@ func testFlush(t *testing.T, client Client) {
 		}
 	}
 	wait()
+	if t.Failed() {
+		return
+	}
 
 	if err := client.Do(ctx, client.B().Flushall().Build()).Error(); err != nil {
 		t.Fatalf("unexpected flush err %v", err)
@@ -85,8 +91,8 @@ func testFlush(t *testing.T, client Client) {
 
 	t.Logf("testing client side caching after flush\n")
 	jobs, wait = parallel(para)
-	for i := 0; i < keys; i++ {
-		key := strconv.Itoa(i)
+	for i := 0; i < keys && !t.Failed(); i++ {
+		key := "f" + strconv.Itoa(i)
 		jobs <- func() {
 			resp := client.DoCache(ctx, client.B().Get().Key(key).Cache(), time.Minute)
 			if !IsRedisNil(resp.Error()) {
@@ -114,15 +120,17 @@ func testSETGET(t *testing.T, client Client, csc bool) {
 	keys := 10000
 	para := 8
 
+	prefix := strconv.Itoa(rand.Intn(100000))
+
 	kvs := make(map[string]string, keys)
 	for i := 0; i < keys; i++ {
-		kvs[strconv.Itoa(i)] = strconv.FormatInt(rand.Int63(), 10)
+		kvs[prefix+strconv.Itoa(i)] = strconv.FormatInt(rand.Int63(), 10)
 	}
 
 	t.Logf("testing SET with %d keys and %d parallelism\n", keys, para)
 	jobs, wait := parallel(para)
-	for i := 0; i < keys; i++ {
-		key := strconv.Itoa(i)
+	for i := 0; i < keys && !t.Failed(); i++ {
+		key := prefix + strconv.Itoa(i)
 		jobs <- func() {
 			val, err := client.Do(ctx, client.B().Set().Key(key).Value(kvs[key]).Build()).ToString()
 			if err != nil || val != "OK" {
@@ -131,11 +139,14 @@ func testSETGET(t *testing.T, client Client, csc bool) {
 		}
 	}
 	wait()
+	if t.Failed() {
+		return
+	}
 
 	t.Logf("testing GET with %d keys and %d parallelism\n", keys*2, para)
 	jobs, wait = parallel(para)
-	for i := 0; i < keys*2; i++ {
-		key := strconv.Itoa(rand.Intn(keys * 2))
+	for i := 0; i < keys*2 && !t.Failed(); i++ {
+		key := prefix + strconv.Itoa(rand.Intn(keys*2))
 		jobs <- func() {
 			val, err := client.Do(ctx, client.B().Get().Key(key).Build()).ToString()
 			if v, ok := kvs[key]; !((ok && val == v) || (!ok && IsRedisNil(err))) {
@@ -144,11 +155,14 @@ func testSETGET(t *testing.T, client Client, csc bool) {
 		}
 	}
 	wait()
+	if t.Failed() {
+		return
+	}
 
 	t.Logf("testing stream GET with %d keys and %d parallelism\n", keys*2, para)
 	jobs, wait = parallel(para)
-	for i := 0; i < keys*2; i++ {
-		key := strconv.Itoa(rand.Intn(keys * 2))
+	for i := 0; i < keys*2 && !t.Failed(); i++ {
+		key := prefix + strconv.Itoa(rand.Intn(keys*2))
 		jobs <- func() {
 			s := client.DoStream(ctx, client.B().Get().Key(key).Build())
 			buf := bytes.NewBuffer(nil)
@@ -159,11 +173,14 @@ func testSETGET(t *testing.T, client Client, csc bool) {
 		}
 	}
 	wait()
+	if t.Failed() {
+		return
+	}
 
 	t.Logf("testing GET with %d keys and %d parallelism with timeout\n", keys*100, para)
 	jobs, wait = parallel(para)
-	for i := 0; i < keys*100; i++ {
-		key := strconv.Itoa(rand.Intn(keys))
+	for i := 0; i < keys*100 && !t.Failed(); i++ {
+		key := prefix + strconv.Itoa(rand.Intn(keys))
 		jobs <- func() {
 			ctx, cancel := context.WithTimeout(context.Background(), time.Millisecond)
 			defer cancel()
@@ -176,12 +193,15 @@ func testSETGET(t *testing.T, client Client, csc bool) {
 		}
 	}
 	wait()
+	if t.Failed() {
+		return
+	}
 
 	t.Logf("testing client side caching with %d interations and %d parallelism\n", keys*5, para)
 	jobs, wait = parallel(para)
 	hits, miss := int64(0), int64(0)
-	for i := 0; i < keys*10; i++ {
-		key := strconv.Itoa(rand.Intn(keys / 100))
+	for i := 0; i < keys*10 && !t.Failed(); i++ {
+		key := prefix + strconv.Itoa(rand.Intn(keys/100))
 		jobs <- func() {
 			resp := client.DoCache(ctx, client.B().Get().Key(key).Cache(), time.Minute)
 			val, err := resp.ToString()
@@ -196,6 +216,9 @@ func testSETGET(t *testing.T, client Client, csc bool) {
 		}
 	}
 	wait()
+	if t.Failed() {
+		return
+	}
 	if csc {
 		if atomic.LoadInt64(&miss) != 100 || atomic.LoadInt64(&hits) != int64(keys*10-100) {
 			t.Fatalf("unexpected client side caching hits and miss %v %v", atomic.LoadInt64(&hits), atomic.LoadInt64(&miss))
@@ -208,8 +231,8 @@ func testSETGET(t *testing.T, client Client, csc bool) {
 
 	t.Logf("testing DEL with %d keys and %d parallelism\n", keys*2, para)
 	jobs, wait = parallel(para)
-	for i := 0; i < keys*2; i++ {
-		key := strconv.Itoa(i)
+	for i := 0; i < keys*2 && !t.Failed(); i++ {
+		key := prefix + strconv.Itoa(i)
 		jobs <- func() {
 			val, err := client.Do(ctx, client.B().Del().Key(key).Build()).AsInt64()
 			if _, ok := kvs[key]; !((val == 1 && ok) || (val == 0 && !ok)) {
@@ -218,13 +241,16 @@ func testSETGET(t *testing.T, client Client, csc bool) {
 		}
 	}
 	wait()
+	if t.Failed() {
+		return
+	}
 
 	time.Sleep(time.Second)
 
 	t.Logf("testing client side caching after delete\n")
 	jobs, wait = parallel(para)
-	for i := 0; i < keys/100; i++ {
-		key := strconv.Itoa(i)
+	for i := 0; i < keys/100 && !t.Failed(); i++ {
+		key := prefix + strconv.Itoa(i)
 		jobs <- func() {
 			resp := client.DoCache(ctx, client.B().Get().Key(key).Cache(), time.Minute)
 			if !IsRedisNil(resp.Error()) {
@@ -268,7 +294,7 @@ func testMultiSETGET(t *testing.T, client Client, csc bool) {
 
 	t.Logf("testing Multi SET with %d keys and %d parallelism\n", keys, para)
 	jobs, wait := parallel(para)
-	for i := 0; i < keys; i += batch {
+	for i := 0; i < keys && !t.Failed(); i += batch {
 		commands := make(Commands, 0, batch)
 		for j := 0; j < batch; j++ {
 			key := "m" + strconv.Itoa(i+j)
@@ -284,10 +310,13 @@ func testMultiSETGET(t *testing.T, client Client, csc bool) {
 		}
 	}
 	wait()
+	if t.Failed() {
+		return
+	}
 
 	t.Logf("testing GET with %d keys and %d parallelism\n", keys*2, para)
 	jobs, wait = parallel(para)
-	for i := 0; i < keys*2; i += batch {
+	for i := 0; i < keys*2 && !t.Failed(); i += batch {
 		cmdkeys := make([]string, 0, batch)
 		commands := make(Commands, 0, batch)
 		for j := 0; j < batch; j++ {
@@ -304,10 +333,13 @@ func testMultiSETGET(t *testing.T, client Client, csc bool) {
 		}
 	}
 	wait()
+	if t.Failed() {
+		return
+	}
 
 	t.Logf("testing GET with %d keys and %d parallelism with timeout\n", keys*100, para)
 	jobs, wait = parallel(para)
-	for i := 0; i < keys*100; i += batch {
+	for i := 0; i < keys*100 && !t.Failed(); i += batch {
 		cmdkeys := make([]string, 0, batch)
 		commands := make(Commands, 0, batch)
 		for j := 0; j < batch; j++ {
@@ -328,11 +360,14 @@ func testMultiSETGET(t *testing.T, client Client, csc bool) {
 		}
 	}
 	wait()
+	if t.Failed() {
+		return
+	}
 
 	t.Logf("testing client side caching with %d interations and %d parallelism\n", keys*5, para)
 	jobs, wait = parallel(para)
 	hits, miss := int64(0), int64(0)
-	for i := 0; i < keys*10; i += batch {
+	for i := 0; i < keys*10 && !t.Failed(); i += batch {
 		cmdkeys := make([]string, 0, batch)
 		commands := make([]CacheableTTL, 0, batch)
 		for j := 0; j < batch; j++ {
@@ -354,6 +389,9 @@ func testMultiSETGET(t *testing.T, client Client, csc bool) {
 		}
 	}
 	wait()
+	if t.Failed() {
+		return
+	}
 	if csc {
 		if atomic.LoadInt64(&miss) != 100 || atomic.LoadInt64(&hits) != int64(keys*10-100) {
 			t.Fatalf("unexpected client side caching hits and miss %v %v", atomic.LoadInt64(&hits), atomic.LoadInt64(&miss))
@@ -366,7 +404,7 @@ func testMultiSETGET(t *testing.T, client Client, csc bool) {
 
 	t.Logf("testing DEL with %d keys and %d parallelism\n", keys*2, para)
 	jobs, wait = parallel(para)
-	for i := 0; i < keys*2; i += batch {
+	for i := 0; i < keys*2 && !t.Failed(); i += batch {
 		cmdkeys := make([]string, 0, batch)
 		commands := make(Commands, 0, batch)
 		for j := 0; j < batch; j++ {
@@ -383,12 +421,15 @@ func testMultiSETGET(t *testing.T, client Client, csc bool) {
 		}
 	}
 	wait()
+	if t.Failed() {
+		return
+	}
 
 	time.Sleep(time.Second)
 
 	t.Logf("testing client side caching after delete\n")
 	jobs, wait = parallel(para)
-	for i := 0; i < keys/100; i += batch {
+	for i := 0; i < keys/100 && !t.Failed(); i += batch {
 		cmdkeys := make([]string, 0, batch)
 		commands := make([]CacheableTTL, 0, batch)
 		for j := 0; j < batch; j++ {
@@ -510,6 +551,9 @@ func testMultiExec(t *testing.T, client Client) {
 	t.Logf("testing MULTI EXEC with %d keys and %d parallelism\n", keys, para)
 	jobs, wait := parallel(para)
 	for k, v := range kvs {
+		if t.Failed() {
+			break
+		}
 		k, v := k, v
 		jobs <- func() {
 			resps, err := client.DoMulti(ctx,
@@ -631,7 +675,7 @@ func testPubSub(t *testing.T, client Client) {
 
 	go func() {
 		time.Sleep(time.Second)
-		for i := 0; i < msgs; i++ {
+		for i := 0; i < msgs && !t.Failed(); i++ {
 			msg := strconv.Itoa(i)
 			ch := "ch1"
 			if i%10 == 0 {
@@ -644,6 +688,9 @@ func testPubSub(t *testing.T, client Client) {
 			}
 		}
 		wait()
+		if t.Failed() {
+			close(messages)
+		}
 	}()
 
 	for message := range messages {
@@ -714,7 +761,7 @@ func testPubSubSharded(t *testing.T, client Client) {
 
 	go func() {
 		time.Sleep(time.Second)
-		for i := 0; i < msgs; i++ {
+		for i := 0; i < msgs && !t.Failed(); i++ {
 			msg := strconv.Itoa(i)
 			ch := "ch1"
 			jobs <- func() {
@@ -724,6 +771,9 @@ func testPubSubSharded(t *testing.T, client Client) {
 			}
 		}
 		wait()
+		if t.Failed() {
+			close(messages)
+		}
 	}()
 
 	for message := range messages {
@@ -771,12 +821,15 @@ func testLua(t *testing.T, client Client) {
 	para := 4
 	kvs := make(map[string]string, keys)
 	for i := 0; i < keys; i++ {
-		kvs["m"+strconv.Itoa(i)] = strconv.FormatInt(rand.Int63(), 10)
+		kvs["l"+strconv.Itoa(i)] = strconv.FormatInt(rand.Int63(), 10)
 	}
 
 	t.Logf("testing lua with %d keys and %d parallelism\n", keys, para)
 	jobs, wait := parallel(para)
 	for k, v := range kvs {
+		if t.Failed() {
+			break
+		}
 		k := k
 		v := v
 		jobs <- func() {
